Se anade la validacion de errrores
-Se excluyen archivos con aristas repetidas
-Se excluyen archivos con vertices repetidas
-Se excluyen archivos con tipos no correspondientes
-Se excluyen archivos con aristas con destino u origen que no esta dentro de las vertices del grafo

+++ src/mrdelivery/model/LectorJSON.java
@@ -1,5 +1,7 @@
 package mrdelivery.model;
 
+import org.json.JSONArray;
+import org.json.JSONException;
 import org.json.JSONObject;
 import java.io.FileReader;
 import java.io.File;
@@ -8,6 +10,8 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
+import java.util.HashSet;
+
 import static java.nio.file.StandardCopyOption.REPLACE_EXISTING;
 
 public class LectorJSON implements Lector {
@@ -17,18 +21,30 @@ public class LectorJSON implements Lector {
     private final String ruta;
     private final String rutaLeidos;
     private boolean lecturaContinua;
+    private final String rutaBorrados;
+
+    public LectorJSON(String rutaRelativa, String _rutaLeidos,String _rutaBorrados){
+        archivo = new File(Paths.get(".",rutaRelativa).toString());
+        archivosJson = new ArrayList<>();
+        ruta = rutaRelativa;
+        rutaLeidos = _rutaLeidos;
+        rutaBorrados = _rutaBorrados;
+        lecturaContinua = true;
+    }
 
     public LectorJSON(String rutaRelativa, String _rutaLeidos){
         archivo = new File(Paths.get(".",rutaRelativa).toString());
         archivosJson = new ArrayList<>();
         ruta = rutaRelativa;
         rutaLeidos = _rutaLeidos;
+        rutaBorrados = null;
         lecturaContinua = true;
     }
     public LectorJSON(String rutaRelativa){
         archivo = new File(Paths.get(".",rutaRelativa).toString());
         archivosJson = new ArrayList<>();
         ruta = rutaRelativa;
+        rutaBorrados = null;
         rutaLeidos = null;
         lecturaContinua = false;
     }
@@ -58,30 +74,42 @@ public class LectorJSON implements Lector {
     }
 
     public void leerArchivo(File archivo) throws IOException {
-        if (archivo.exists() && archivo.toString().endsWith(".json")){
+        if (archivo.exists() && archivo.toString().endsWith(".json")) {
             FileReader lector = new FileReader(archivo);
             StringBuilder datos = new StringBuilder();
             int caracterLeido = 0;
-            while ((caracterLeido = lector.read()) != -1){
-                datos.append((char)caracterLeido);
+            while ((caracterLeido = lector.read()) != -1) {
+                datos.append((char) caracterLeido);
             }
             JSONObject json = new JSONObject(datos.toString());
-            archivosJson.add(json);
-            System.out.println("Leyo: " + archivo.getName());
-            lector.close();
-            // Despues de haber leido el archivo, se mueve a la carpeta de leidos si es un lector continuo
-            if (lecturaContinua) {
-                String nuevaUbicacion = String.valueOf(Paths.get("",rutaLeidos.substring(1,rutaLeidos.length())+"/"+archivo.getName()));
-                try {
-                    Files.copy(archivo.toPath(), Path.of(nuevaUbicacion),REPLACE_EXISTING);
+                archivosJson.add(json);
+                System.out.println("Leyo: " + archivo.getName());
+                lector.close();
+                if (lecturaContinua) {
+                    if (validarJSON(json)) {
+                    String nuevaUbicacion = String.valueOf(Paths.get("", rutaLeidos.substring(1, rutaLeidos.length()) + "/" + archivo.getName()));
+                    try {
+                        Files.copy(archivo.toPath(), Path.of(nuevaUbicacion), REPLACE_EXISTING);
 
-                    Files.delete(archivo.toPath());
+                        Files.delete(archivo.toPath());
 
+                    } catch (Exception e) {
+                        System.out.println(e.toString());
+                        System.out.println("No se pudo mandar el archivo a los leidos");
+                    }
                 }
-                catch (Exception e){
-                    System.out.println(e.toString());
-                    System.out.println("No se pudo mandar el archivo a los leidos");
-                }
+                    else {
+                        String nuevaUbicacion = String.valueOf(Paths.get("", rutaBorrados.substring(1, rutaBorrados.length()) + "/" + archivo.getName()));
+                        try {
+                            Files.copy(archivo.toPath(), Path.of(nuevaUbicacion), REPLACE_EXISTING);
+                            Files.delete(archivo.toPath());
+                        }
+                        catch (Exception e){
+                            System.out.println(e.toString());
+                            System.out.println("No se pudo mandar el archivo a los errores");
+                        }
+                        System.out.println("Archivo con errores");
+                    }
             }
         }
         else
@@ -99,6 +127,8 @@ public class LectorJSON implements Lector {
             System.out.println("El fichero " + fichero.getAbsolutePath() + " no existe");
     }
 
+
+
     public ArrayList<JSONObject> getArchivosJson(){
         return archivosJson;
     }
@@ -106,4 +136,57 @@ public class LectorJSON implements Lector {
     public String obtenerRutaJSON(){
         return ruta;
     }
+
+    public boolean validarJSON(JSONObject json){
+        System.out.println(json);
+        JSONArray vertices = json.getJSONArray("vertices");
+        JSONArray aristas = json.getJSONArray("aristas");
+        HashSet<String > nombresVertices = new HashSet<>();
+        try {
+            for (int i = 0;i < vertices.length();i++){
+                String nombre = vertices.getString(i);
+                nombresVertices.add(nombre);
+            }
+        }
+        catch (JSONException e){
+            System.out.println("ERROR NOMBRE DE VERTICE NO ES UN STRING");
+            return false;
+        }
+        if(nombresVertices.size() != vertices.length()) {
+            System.out.println("ERROR NOMBRE DE VERTICE REPETIDO");
+            return false;
+        }
+        else
+            return validarTodasAristas(nombresVertices,aristas);
+    }
+
+    public boolean validarTodasAristas(HashSet<String> vertices,JSONArray aristas){//TODO:CONSIDERAR SI VIENE VACIO O UN ELEMENTO
+        HashSet<String> origenDestino = new HashSet<>();
+        for (int i = 0;i< aristas.length();i++){
+            try {
+                JSONObject arista = aristas.getJSONObject(i);
+                String origen = arista.getString("origen");
+                String destino = arista.getString("destino");
+                arista.getBoolean("activo");
+                arista.getDouble("costo");
+                arista.getDouble("km");
+                arista.getDouble("minutos");
+                origenDestino.add(origen+destino);
+                if(!(vertices.contains(origen) && vertices.contains(destino))) {
+                    System.out.println("NO EXISTE EL VERTICE PARA ASIGNAR LA ARISTA");
+                    return false;
+                }
+            }
+            catch (JSONException e) {
+                System.out.println("LOS VALORES DE LAS ARISTAS NO SON CORRECTOS");
+                return false;
+            }
+        }
+        if(origenDestino.size() != aristas.length()) {
+            System.out.println("ERROR ARISTA REPETIDO");
+            return false;
+        }
+        return true;
+    }
+
 }
